app/bot/requirements.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
autopep8==1.5.5
Babel==2.9.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==2.1.1
distlib==0.3.6
filelock==3.9.0
frozenlist==1.3.3
idna==2.10
magic-filter==1.0.9
multidict==6.0.4
mysql-connector-python==8.0.29
protobuf==4.21.12
pycodestyle==2.10.0
PyJWT==2.6.0
pyTelegramBotAPI==3.7.9
python-dateutil==2.8.1
python-dotenv==1.0.0
python-environ==0.4.54
pytz==2022.7.1
redis==4.5.1
requests==2.25.1
six==1.16.0
toml==0.10.2
urllib3==1.26.14
virtualenv==20.4.0
yarl==1.8.2
